Moved all json to a reference folder
