+++ src/reference/monsters.json
